moved rrdcalc and rrdvar to health
