+++ database/rrd.h
@@ -103,8 +103,8 @@ struct ml_metrics_statistics {
 #include "daemon/common.h"
 #include "web/api/queries/query.h"
 #include "web/api/queries/rrdr.h"
-#include "rrdvar.h"
-#include "rrdcalc.h"
+#include "health/rrdvar.h"
+#include "health/rrdcalc.h"
 #include "rrdlabels.h"
 #include "streaming/rrdpush.h"
 #include "aclk/aclk_rrdhost_state.h"
